Add inbound common model class
